requirements.txt
Django==1.8.1
psycopg2==2.6
gunicorn==19.3.0
whitenoise==1.0.6
django-debug-toolbar==1.3.0
django-storages-redux
py-trello
django-cors-headers
google-api-python-client
gspread

